Move event filtering classes up to the package level
